Create a constructRoomsItems for building the rooms, items, and inventory.
By moving this out of main and putting it in its own method,
we can run it each time we start a new game to reset things.
We need this because we need to reset rooms and inventory each new game.
We could technically construct the items outside this method
because they don't change but I think
doing all of this together helps readability.