Check if the player has the ingredients when shooting the garage door.
If they do, they lose the game.
The problem is that the user could shoot out the garage door without
giving ingredients to Eleanor.
They could then get the hallway key and make the game go on.
This causes a problem because then Eleanor mentions having the antidote
but player still has ingredients which doesn't make sense.
When shooting the door  make the player drop the ingredients which
gives them a gameover and hence solves this problem.